Change parameter type of Amount from int to float

+++ db/init.sql
@@ -1,8 +1,7 @@
--- Table Definition
 CREATE TABLE IF NOT EXISTS expenses (
 		id SERIAL PRIMARY KEY,
 		title TEXT,
-		amount int,
+		amount float,
 		note TEXT,
 		tags TEXT[]
 );
